Move uuid enums to ~/model/decoder
