+++ __fixtures__/uuid/article.ts
@@ -24,10 +24,7 @@ import * as R from 'ramda'
 import { license } from '../license'
 import { getRepositoryDataWithoutSubResolvers } from './abstract-repository'
 import { Model } from '~/internals/graphql'
-import {
-  EntityRevisionType,
-  EntityType,
-} from '~/schema/uuid/abstract-entity/types'
+import { EntityRevisionType, EntityType } from '~/model/decoder'
 import { Instance } from '~/types'
 
 export const article: Model<'Article'> = {
