+++ __tests__/schema/uuid/abstract-taxonomy-term-child.ts
@@ -40,7 +40,7 @@ import {
   createUuidHandler,
 } from '../../__utils__'
 import { Model } from '~/internals/graphql'
-import { EntityType } from '~/schema/uuid/abstract-entity/types'
+import { EntityType } from '~/model/decoder'
 
 let client: Client
