+++ src/api.ts
@@ -36,6 +36,5 @@ export * from '~/model/types'
 export * from '~/schema/connection/types'
 export * from '~/schema/subscription/types'
 export * from '~/schema/uuid/abstract-navigation-child/types'
-export * from '~/schema/uuid/abstract-repository/types'
 export * from '~/types'
 export * from '~/utils'
