+++ src/model/decoder.ts
@@ -21,15 +21,7 @@
  */
 import * as t from 'io-ts'
 
-import {
-  EntityRevisionType,
-  EntityType,
-} from '~/schema/uuid/abstract-entity/types'
-import { DiscriminatorType } from '~/schema/uuid/abstract-uuid/types'
 import { Instance, TaxonomyTermType } from '~/types'
-
-export * from '~/schema/uuid/abstract-entity/types'
-export * from '~/schema/uuid/abstract-uuid/types'
 
 export const InstanceDecoder: t.Type<Instance> = t.union([
   t.literal(Instance.De),
@@ -39,6 +31,45 @@ export const InstanceDecoder: t.Type<Instance> = t.union([
   t.literal(Instance.Hi),
   t.literal(Instance.Ta),
 ])
+
+export enum DiscriminatorType {
+  Comment = 'Comment',
+  Page = 'Page',
+  PageRevision = 'PageRevision',
+  TaxonomyTerm = 'TaxonomyTerm',
+  User = 'User',
+}
+
+export type UuidType = DiscriminatorType | EntityType | EntityRevisionType
+
+export type RepositoryType = EntityType | DiscriminatorType.Page
+export type RevisionType = EntityRevisionType | DiscriminatorType.PageRevision
+
+export enum EntityType {
+  Applet = 'Applet',
+  Article = 'Article',
+  Course = 'Course',
+  CoursePage = 'CoursePage',
+  Event = 'Event',
+  Exercise = 'Exercise',
+  ExerciseGroup = 'ExerciseGroup',
+  GroupedExercise = 'GroupedExercise',
+  Solution = 'Solution',
+  Video = 'Video',
+}
+
+export enum EntityRevisionType {
+  ArticleRevision = 'ArticleRevision',
+  AppletRevision = 'AppletRevision',
+  CourseRevision = 'CourseRevision',
+  CoursePageRevision = 'CoursePageRevision',
+  EventRevision = 'EventRevision',
+  ExerciseRevision = 'ExerciseRevision',
+  ExerciseGroupRevision = 'ExerciseGroupRevision',
+  GroupedExerciseRevision = 'GroupedExerciseRevision',
+  SolutionRevision = 'SolutionRevision',
+  VideoRevision = 'VideoRevision',
+}
 
 // As of 26.03.2021 the maximum uuid is 201517. Thus there are ~200.000 uuids
 // per 10 years. The following maximum shouldn't be hit in the next ~ 40 years.
