+++ src/model/serlo.ts
@@ -29,13 +29,16 @@ import {
   InstanceDecoder,
   NotificationEventDecoder,
   UuidDecoder,
-  // TODO: The following import is needed for the API extractor
-  // Delete the line when https://github.com/microsoft/rushstack/issues/2140
-  // got fixed
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  NotificationEventType,
   Uuid,
   NotificationDecoder,
+  // TODO: The following imports are needed for the API extractor
+  // Delete them when https://github.com/microsoft/rushstack/issues/2140 is fixed
+  /* eslint-disable @typescript-eslint/no-unused-vars */
+  NotificationEventType,
+  EntityType,
+  EntityRevisionType,
+  DiscriminatorType,
+  /* eslint-enable @typescript-eslint/no-unused-vars */
 } from './decoder'
 import { Environment } from '~/internals/environment'
 import { Model } from '~/internals/graphql'
@@ -56,7 +59,12 @@ import {
 } from '~/schema/uuid/abstract-navigation-child/types'
 import { isUnsupportedUuid } from '~/schema/uuid/abstract-uuid/utils'
 import { decodePath, encodePath } from '~/schema/uuid/alias/utils'
-import { Instance, ThreadCreateThreadInput } from '~/types'
+import {
+  Instance,
+  ThreadCreateThreadInput,
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  TaxonomyTermType,
+} from '~/types'
 
 export function createSerloModel({
   environment,
